Clock #5 - Fixed spaces between rectangels
I am checking if there is a rectangle to the right or bottom and increasing the width or height by 1 to 'connect' them. It is a somewhat ugly solution since it is now always drawing every rectangle twice (to prevent an extra 1 by 1 rectangle when both height and width are increased on the same rectangle). Maybe it would be better to draw an extra 1 by 'size' rectangle at the bottom and/or right?