Removed gaps from clock 12
by creating a font shape file. It contains an object that defines for
every character the necessary shapes. It also contains a draw method
that needs a p5-object and the character. By setting the transformation
and the stroke and fill values you can manipulate the look of the
character.